Possible future bug: tables are now Point instead of qglviwer::Vec to be aligned in memory (isomorph to double[3]). Thus, passing stride = 0 to glVertexPointer (0 means contiguous). Bug: index array is 32-bit unsigned int (GLuint) because of OpenGL's limitations. Beware of indexing a DTM with more than 4,294,967,295 vertices... Filtering OpenGL errors at the end of draw_data.

+++ src/dtm.h
@@ -48,10 +48,10 @@ private:
 
 protected:
     template<Primitive prim> inline void build_begin() const;
-    template<Primitive prim> inline void build_line(unsigned long);
-    template<Primitive prim> inline void build_line(unsigned long) const;
-    template<Primitive prim> inline void build_back(unsigned long);
-    template<Primitive prim> inline void build_back(unsigned long) const;
+    template<Primitive prim> inline void build_line(GLuint);
+    template<Primitive prim> inline void build_line(GLuint) const;
+    template<Primitive prim> inline void build_back(GLuint);
+    template<Primitive prim> inline void build_back(GLuint) const;
     template<Primitive prim> inline void build_end() const;
 
     inline unsigned long array_size(Primitive prim) const;
